Major code refactor
